+++ EventPlanner/app/src/main/java/com/example/eventplanner/adapters/AllEventsAdapter.java
@@ -13,7 +13,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.eventplanner.EventDetailsActivity;
+import com.example.eventplanner.activities.EventDetailsActivity;
 import com.example.eventplanner.R;
 import com.example.eventplanner.models.Event;
