+++ EventPlanner/app/src/main/java/com/example/eventplanner/fragments/AllProductsServicesFragment.java
@@ -1,4 +1,4 @@
-package com.example.eventplanner;
+package com.example.eventplanner.fragments;
 
 import android.app.DatePickerDialog;
 import android.os.Bundle;
@@ -16,18 +16,20 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.eventplanner.adapters.AllEventsAdapter;
-import com.example.eventplanner.models.Event;
+import com.example.eventplanner.R;
+import com.example.eventplanner.adapters.AllProductsServicesAdapter;
+import com.example.eventplanner.models.ProductService;
 
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 
-public class AllEventsFragment extends Fragment {
+public class AllProductsServicesFragment extends Fragment {
 
-    private AllEventsAdapter allEventsAdapter;
-    private List<Event> allEventsList;
+    private AllProductsServicesAdapter allProductsServicesAdapter;
+    private List<ProductService> allProductsServicesList;
     private EditText searchEditText;
+
     private Button fromDateButton, toDateButton;
     private Calendar calendar;
 
@@ -35,28 +37,28 @@ public class AllEventsFragment extends Fragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        allEventsList = new ArrayList<>();
-        allEventsList.add(new Event("Event 1", "Description for Event 1", R.drawable.event_placeholder4));
-        allEventsList.add(new Event("Event 2", "Description for Event 2", R.drawable.event_placeholder));
-        allEventsList.add(new Event("Event 3", "Description for Event 3", R.drawable.event_placeholder4));
-        allEventsList.add(new Event("Event 4", "Description for Event 4", R.drawable.event_placeholder));
-        allEventsList.add(new Event("Event 5", "Description for Event 5", R.drawable.event_placeholder));
-        allEventsList.add(new Event("Event 1", "Description for Event 1", R.drawable.event_placeholder2));
-        allEventsList.add(new Event("Event 2", "Description for Event 2", R.drawable.event_placeholder2));
-        allEventsList.add(new Event("Event 3", "Description for Event 3", R.drawable.event_placeholder2));
-        allEventsList.add(new Event("Event 4", "Description for Event 4", R.drawable.event_placeholder));
-        allEventsList.add(new Event("Event 5", "Description for Event 5", R.drawable.event_placeholder));
+        allProductsServicesList = new ArrayList<>();
+        allProductsServicesList.add(new ProductService("Product 1", "Description for Product 1", R.drawable.product_service_placeholder2));
+        allProductsServicesList.add(new ProductService("Service 1", "Description for Service 1", R.drawable.product_service_placeholder3));
+        allProductsServicesList.add(new ProductService("Product 2", "Description for Product 2", R.drawable.product_service_placeholder3));
+        allProductsServicesList.add(new ProductService("Product 3", "Description for Product 3", R.drawable.product_service_placeholder));
+        allProductsServicesList.add(new ProductService("Service 2", "Description for Service 2", R.drawable.product_service_placeholder4));
+        allProductsServicesList.add(new ProductService("Product 4", "Description for Product 4", R.drawable.product_service_placeholder4));
+        allProductsServicesList.add(new ProductService("Product 5", "Description for Product 5", R.drawable.product_service_placeholder4));
+        allProductsServicesList.add(new ProductService("Service 3", "Description for Service 3", R.drawable.product_service_placeholder));
+        allProductsServicesList.add(new ProductService("Product 6", "Description for Product 6", R.drawable.product_service_placeholder));
+        allProductsServicesList.add(new ProductService("Product 7", "Description for Product 7", R.drawable.product_service_placeholder3));
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.all_events_section, container, false);
+        View view = inflater.inflate(R.layout.all_products_services_section, container, false);
 
-        RecyclerView allEventsRecyclerView = view.findViewById(R.id.all_events_recycler_view);
-        allEventsRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
-        allEventsAdapter = new AllEventsAdapter(getContext(), allEventsList);
-        allEventsRecyclerView.setAdapter(allEventsAdapter);
+        RecyclerView allProductsServicesRecyclerView = view.findViewById(R.id.all_products_services_recycler_view);
+        allProductsServicesRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
+        allProductsServicesAdapter = new AllProductsServicesAdapter(getContext(), allProductsServicesList);
+        allProductsServicesRecyclerView.setAdapter(allProductsServicesAdapter);
 
         Spinner sortSpinner = view.findViewById(R.id.sort_spinner);
         Spinner filterSpinner = view.findViewById(R.id.filter_spinner);
@@ -73,11 +75,11 @@ public class AllEventsFragment extends Fragment {
         filterAdapter.setDropDownViewResource(R.layout.spinner_item);
         filterSpinner.setAdapter(filterAdapter);
 
-        searchIcon.setOnClickListener(v -> filterEvents(searchEditText.getText().toString()));
+        searchIcon.setOnClickListener(v -> filterProductsServices(searchEditText.getText().toString()));
         sortSpinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
             public void onItemSelected(AdapterView<?> parentView, View view, int position, long id) {
-                sortEvents(position);
+                sortProductsServices(position);
             }
 
             @Override
@@ -87,7 +89,7 @@ public class AllEventsFragment extends Fragment {
         filterSpinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
             public void onItemSelected(AdapterView<?> parentView, View view, int position, long id) {
-                filterEventsByType(position);
+                filterProductsServicesByType(position);
             }
 
             @Override
@@ -120,22 +122,22 @@ public class AllEventsFragment extends Fragment {
         datePickerDialog.show();
     }
 
-    private void filterEvents(String query) {
-        List<Event> filteredEvents = new ArrayList<>();
-        for (Event event : allEventsList) {
-            if (event.getTitle().toLowerCase().contains(query.toLowerCase()) ||
-                    event.getDescription().toLowerCase().contains(query.toLowerCase())) {
-                filteredEvents.add(event);
+    private void filterProductsServices(String query) {
+        List<ProductService> filteredProductsServices = new ArrayList<>();
+        for (ProductService productService : allProductsServicesList) {
+            if (productService.getTitle().toLowerCase().contains(query.toLowerCase()) ||
+                    productService.getDescription().toLowerCase().contains(query.toLowerCase())) {
+                filteredProductsServices.add(productService);
             }
         }
-        allEventsAdapter.updateEventList(filteredEvents);
+        allProductsServicesAdapter.updateProductServiceList(filteredProductsServices);
     }
 
-    private void sortEvents(int sortType) {
-        // Implement sorting logic for events
+    private void sortProductsServices(int sortType) {
+        // Implement sorting logic for products/services
     }
 
-    private void filterEventsByType(int filterType) {
-        // Implement filtering logic for events
+    private void filterProductsServicesByType(int filterType) {
+        // Implement filtering logic for products/services
     }
 }
