+++ EventPlanner/app/src/main/java/com/example/eventplanner/models/ServiceCategory.java
@@ -1,4 +1,4 @@
-package com.example.eventplanner;
+package com.example.eventplanner.models;
 
 public class ServiceCategory {
     private String name;
