+++ EventPlanner/app/src/main/res/layout/activity_all_services.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".AllServicesActivity">
+    tools:context=".activities.AllServicesActivity">
 
     <TextView
         android:id="@+id/Title"
