+++ EventPlanner/app/src/main/res/layout/activity_splash.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".activities.MainActivity"
     android:background="@drawable/ic_launcher_background">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
