+++ EventPlanner/app/src/main/res/layout/top_5_products_services_carousel.xml
@@ -4,8 +4,7 @@
     android:orientation="vertical"
     android:gravity="center_horizontal"
     android:padding="16dp"
-    android:background="@android:color/white"
-    android:layout_marginBottom="16dp">
+    android:background="@android:color/white">
 
     <TextView
         android:id="@+id/carousel_label"
