+++ EventPlanner/app/src/main/res/menu/navigation_menu.xml
@@ -1,23 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/nav_home"
-        android:title="@string/home"/>
-    <item android:id="@+id/nav_chat"
-        android:title="@string/chat"/>
-    <item android:id="@+id/nav_notifications"
-        android:title="@string/notifications"/>
-    <item android:id="@+id/nav_events"
-        android:title="@string/events"/>
-    <item android:id="@+id/nav_services"
-        android:title="@string/services"/>
-    <item android:id="@+id/nav_products"
-        android:title="@string/products"/>
-    <item android:id="@+id/nav_profile"
-        android:title="@string/profile"/>
-    <item android:id="@+id/nav_settings"
-        android:title="@string/settings"/>
-    <item android:id="@+id/nav_about"
-        android:title="@string/about"/>
-    <item android:id="@+id/nav_sign_out"
-        android:title="@string/sign_out"/>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:showIn="navigation_view">
+        <group android:checkableBehavior="single">
+                <item android:id="@+id/nav_home"
+                    android:title="@string/home"
+                    android:icon="@drawable/ic_home" />
+
+                <item android:id="@+id/nav_chat"
+                    android:title="@string/chat"
+                    android:icon="@drawable/ic_chat" />
+
+                <item android:id="@+id/nav_notifications"
+                    android:title="@string/notifications"
+                    android:icon="@drawable/ic_notification" />
+
+                <item android:id="@+id/nav_events"
+                    android:title="@string/events"
+                    android:icon="@drawable/ic_event" />
+
+                <item android:id="@+id/nav_services"
+                    android:title="@string/services"
+                    android:icon="@drawable/ic_suitcase" />
+
+                <item android:id="@+id/nav_products"
+                    android:title="@string/products"
+                    android:icon="@drawable/ic_box" />
+
+                <item android:id="@+id/nav_profile"
+                    android:title="@string/profile"
+                    android:icon="@drawable/ic_user" />
+
+                <item android:id="@+id/nav_sign_out"
+                    android:title="@string/sign_out"
+                    android:icon="@drawable/ic_sign_out" />
+        </group>
 </menu>
