+++ EventPlanner/app/src/main/res/values/styles.xml
@@ -4,4 +4,7 @@
         <item name="cornerFamily">rounded</item>
         <item name="cornerSize">16dp</item>
     </style>
+    <style name="NavigationDrawerTextStyle">
+        <item name="android:fontFamily">@font/raleway</item>
+    </style>
 </resources>
